Add build integration into Github Actions using docker compose (#11)
* Docker compose

* Build docker images and upload in action

* Add dockerignore

* Remove build configuration

I was unable to build the server image using this setting, I'm not sure
why.

* Update lock file

* Revert generating openapi specification in docker

This does not fully work yet and probably is over complicated for our
needs.

* Fix dev compose

* Remove PUBLIC_API_BASE_URL as it's not being used

Before we would use this to tell the UI which host to use as the api
host. But as this configuration serves both the UI and server on the
same domain (by levaraging traefik) we don't need this.

* Add dpepends_on

* Add config.yaml to Dockerfile

* Fix ui watch path

* Set base url to be empty by default

Before having `/` as the default value would mean we sometimes end up
with a double forward slash `//`.