Improve web rtc integration (#38)
* Improve webrtc ux & connection logic

We've now moved the logic of connecting to webrtc into a separate class
that can be used outside of react, which allows us to have better
control over status changes and error handling.

With these changes we also add a loading indicator when we start to
connect to the webrtc server.

* Add debug trigger to help debug during development

* Use spectrum aliases and don't immediately stop on error